change datatype phone number

+++ app/models/user.js
@@ -55,7 +55,7 @@ module.exports = (sequelize, DataTypes) => {
       lastName: DataTypes.STRING,
       NIK: DataTypes.BIGINT,
       address: DataTypes.STRING,
-      phoneNumber: DataTypes.INTEGER,
+      phoneNumber: DataTypes.BIGINT,
       image: DataTypes.TEXT,
       dateOfBirth: DataTypes.DATEONLY,
       gender: DataTypes.ENUM("Male", "Female"),
